+++ domain/memory/piastres/expenses/sdk.go
@@ -78,7 +78,7 @@ type Expense interface {
 // ContentBuilder represents a content builder
 type ContentBuilder interface {
 	Create() ContentBuilder
-	WithAmount(amount uint) ContentBuilder
+	WithAmount(amount uint64) ContentBuilder
 	From(from bills.Bill) ContentBuilder
 	WithCancel(cancel locks.Lock) ContentBuilder
 	WithRemaining(remaining locks.Lock) ContentBuilder
@@ -89,7 +89,7 @@ type ContentBuilder interface {
 // Content represents an expense content
 type Content interface {
 	entities.Immutable
-	Amount() uint
+	Amount() uint64
 	From() bills.Bill
 	Cancel() locks.Lock
 	HasRemaining() bool
