+++ domain/transfers/piastres/bills/json_bill.go
@@ -5,7 +5,7 @@ import "time"
 type jsonBill struct {
 	Hash      string    `json:"hash"`
 	Lock      string    `json:"lock"`
-	Amount    uint      `json:"amount"`
+	Amount    uint64    `json:"amount"`
 	CreatedOn time.Time `json:"created_on"`
 }
 
@@ -20,7 +20,7 @@ func createJSONBillFromBill(ins Bill) *jsonBill {
 func createJSONBill(
 	hash string,
 	lock string,
-	amount uint,
+	amount uint64,
 	createdOn time.Time,
 ) *jsonBill {
 	out := jsonBill{
